backend-code-contract-generator/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
pyannote.audio==3.1.1
torch
torchaudio
numpy<2.0
langchain-core
langchain-ollama
langchain-community
sentence-transformers
pydantic
chromadb
pypdf
jinja2
reportlab
python-dotenv